Remove unused fields from the frame header
The frame sync word ("MAGIC_BYTE") still exists, but it's the modem's responsibility, not part of the Golay-coded header.

The frame type word is superceded by the UDP destination port (for UDP packets).

We end up with three whole bytes of reserved header. That's a lot, but if we need even one more bit, that's what it will take to fill out an integer number of 3-byte Golay code plaintexts.  Alternate plan would be to steal a bit from the token, and use it to indicate whether the remaining 23 bits are the token or reserved extension bits. But the numerology is currently set up for 12 bit header, so we will stick with that for now.